requirements.txt
addict==2.4.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
chumpy==0.70
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.1.0
cycler==0.11.0
Cython==0.29.33
einops==0.6.0
exceptiongroup==1.1.0
flatbuffers==23.1.21
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
json-tricks==3.16.1
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
markdown-it-py==2.1.0
matplotlib==3.5.3
mdurl==0.1.2
mmcv-full==1.5.0
mmdet==2.28.1
# Editable install with no version control (mmpose==0.24.0)
-e f:\vitpose\vitpose
model-index==0.1.11
mpmath==1.2.1
munkres==1.1.4
ninja==1.11.1
numpy==1.21.6
onnx==1.12.0
onnxoptimizer==0.3.7
onnxruntime==1.13.1
opencv-python==4.7.0.68
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.1.5
Pillow==9.4.0
pluggy==1.0.0
progressbar2==4.2.0
protobuf==3.19.6
psutil==5.9.4
pycocotools==2.0.6
Pygments==2.14.0
pyparsing==3.0.9
pyreadline==2.1
pytest==7.2.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-utils==3.5.2
PyTurboJPEG==1.7.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scipy==1.7.3
six==1.16.0
sympy==1.10.1
tabulate==0.9.0
terminaltables==3.1.10
tifffile==2021.11.2
timm==0.4.9
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
typing_extensions==4.4.0
urllib3==1.26.14
xtcocotools==1.13
yapf==0.32.0
zipp==3.12.1
